Unwrap snippets in vscode-go
vscode-go snippets are all nested under a ".source.go" key. I can't find
any mention of this type of formatting in the vscode docs, but it
obviously functions fine there. Vsnip will also parse these snippets,
but luasnip will not. Easiest fix is to just normalize them to be the
format luasnip expects.